added gameLogic.h
